pyproject.toml
[tool.poetry]
name = "beard_api"
version = "0.1.0"
description = ""
authors = ["bekzod-fayzikuloff <fayzikuloff.b@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.3"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
djangorestframework = "^3.13.1"
Pillow = "^9.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
